feat: add notification about socket params
