+++ index.ts
@@ -1,6 +1,11 @@
 import { handleData } from './src/websocket/handleData';
 import { WebSocketServer, WebSocket } from 'ws';
-import { users, rooms } from './src/websocket/db';
+import {
+  connectionCount,
+  setCommectionsCout,
+  users
+} from './src/websocket/variables';
+import { rooms } from './src/websocket/db';
 import { httpServer } from './src/http_server/index';
 import { Connection } from './src/models';
 import { handleUsers } from './src/websocket/handleUsers';
@@ -9,10 +14,16 @@ let connections: Connection[] = [];
 export const wss = new WebSocketServer({ server: httpServer });
 
 wss.on('connection', function connection(ws: WebSocket & { userID: number }) {
+  setCommectionsCout(connectionCount + 1);
+  console.log(
+    `Websocket connection established succesfully.\nCurrently there are ${connectionCount} ws connections`
+  );
+
   ws.on('error', console.error);
 
   ws.on('message', function message(data: string) {
-    // console.log(`Received message ${data}`);
+    console.log(`Received message ${data}`);
+
     const parsedData: any = JSON.parse(data);
     if (parsedData.type === 'reg') {
       const innerData = handleUsers(parsedData);
@@ -50,6 +61,7 @@ wss.on('connection', function connection(ws: WebSocket & { userID: number }) {
   ws.on('close', function () {
     rooms.length = 0;
     users.length = 0;
+    setCommectionsCout(connectionCount - 1);
     console.log('connection closed');
   });
 });
