+++ src/websocket/db.ts
@@ -1,23 +1,4 @@
-import { Room, User, ShipsData, Game, Connection } from '../models';
+import { Room, User } from '../models';
 
-export const users: User[] = [];
+export const usersData: User[] = [];
 export const rooms: Room[] = [];
-export const placedShips: ShipsData[] = [];
-export const usersCountInTheRoom = [];
-export let connections: Connection[] = [];
-export let startGame = false;
-export const games: Game[] = [];
-export const turns: number[] = [];
-export let turnUserId: number;
-export const allShipsData: {
-  ownerId: number;
-  ships: any[];
-}[] = [];
-export let currentShootStatus = 'missed';
-
-export function setCurrentShootStatus(value: string) {
-  currentShootStatus = value;
-}
-export function setTurnUserId(value: number) {
-  turnUserId = value;
-}
