+++ src/websocket/handleAttacks.ts
@@ -1,4 +1,4 @@
-import { placedShips, turns } from './db';
+import { placedShips, turns } from './variables';
 import { handleShootStatus } from './handleShootStatus';
 
 export function handleAttaks(parsedData: any, userID: number) {
