+++ src/websocket/handleUsers.ts
@@ -1,15 +1,21 @@
-import { users, connections } from './db';
+import { users } from './variables';
 import { UserData } from '../models';
+import { usersData } from './db';
 
 export function handleUsers(parsedData: UserData) {
   const userName = JSON.parse(parsedData.data);
-  const userDB = {
+  const user = {
     name: JSON.stringify(userName.name),
     index: users.length + 1
   };
+  const userDB = {
+    name: JSON.stringify(userName.name),
+    password: JSON.stringify(userName.password),
+    index: users.length + 1
+  };
 
-  users.push(userDB);
-
+  users.push(user);
+  usersData.push(userDB);
   const innerData = {
     name: userDB.name,
     index: userDB.index,
