+++ src/websocket/updateRooms.ts
@@ -1,4 +1,5 @@
-import { rooms, users } from './db';
+import { users } from './variables';
+import { rooms } from './db';
 
 export function updateExistingRooms(parsedData: any, userId: number) {
   const user = users.find((user) => user.index === userId);
